Comparisons and some optimizations
Optimizations:
* Use 63 bit morton codes for BVH construction instead of 30 Bit.

Comparisons:
* IGL only triangles (original fast winding numbers paper)
* Brute Force (own)
* Brute Force (torch)
* Fast Dipol Sums paper (only point normals)

+++ src/cpp/kernels/build_binary_tree.cuh
@@ -19,7 +19,7 @@ void gather_triangles_soa(const float *__restrict__ input_triangles,
                           float *__restrict__ output_triangles_soa,
                           uint32_t count, const cudaStream_t &stream = 0);
 
-void build_binary_topology(const uint32_t *__restrict__ morton_codes,
+void build_binary_topology(const uint64_t *__restrict__ morton_codes,
                            BinaryNode *nodes, uint32_t *parents,
                            uint32_t leaf_count, const cudaStream_t &stream = 0);
